Changed <int>marks into <size_t>marks

+++ src/FastaReader.cpp
@@ -1,6 +1,6 @@
 #include "FastaReader.h"
 
-std::vector<std::string> FastaReader::readFile(std::vector<int> marks, const std::string file_to_read) 
+std::vector<std::string> FastaReader::readFile(std::vector<size_t> marks, const std::string file_to_read) 
 {
     std::vector<std::string> all_combinations;
     try {
@@ -42,7 +42,7 @@ std::vector<std::string> FastaReader::readFile(std::vector<int> marks, const std
     }
 }
 
-std::map<std::string, double> FastaReader::retrieveData(std::vector<int> marks, const std::string file_to_read) 
+std::map<std::string, double> FastaReader::retrieveData(std::vector<size_t> marks, const std::string file_to_read) 
 {
     std::map<std::string, double> alleles_freq;
     std::vector<std::string> all_combinations_ = readFile(marks, file_to_read);
@@ -55,7 +55,7 @@ std::map<std::string, double> FastaReader::retrieveData(std::vector<int> marks, 
     return alleles_freq;
 }
 
-size_t FastaReader::size(std::vector<int> marks, const std::string file_to_read) 
+size_t FastaReader::size(std::vector<size_t> marks, const std::string file_to_read) 
 {
     return readFile(marks, file_to_read).size();
 }
